PP-4126 Refactor ChargesApiResourceITest
- Refactor ChargesApiResourceITest into more mmanageable chunks
