+++ src/test/java/uk/gov/pay/connector/it/resources/ChargesApiCreateResourceITest.java
@@ -1,0 +1,316 @@
+package uk.gov.pay.connector.it.resources;
+
+import com.google.common.collect.ImmutableMap;
+import com.jayway.restassured.response.ValidatableResponse;
+import org.junit.Test;
+import uk.gov.pay.connector.it.base.ChargingITestBase;
+
+import javax.ws.rs.core.Response.Status;
+import java.util.HashMap;
+
+import static com.jayway.restassured.http.ContentType.JSON;
+import static java.time.temporal.ChronoUnit.SECONDS;
+import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
+import static javax.ws.rs.core.Response.Status.NOT_FOUND;
+import static javax.ws.rs.core.Response.Status.OK;
+import static org.apache.commons.lang3.RandomStringUtils.randomAlphabetic;
+import static org.apache.commons.lang3.RandomStringUtils.randomAlphanumeric;
+import static org.hamcrest.Matchers.hasSize;
+import static org.hamcrest.Matchers.notNullValue;
+import static org.hamcrest.Matchers.nullValue;
+import static org.hamcrest.core.Is.is;
+import static org.hamcrest.text.MatchesPattern.matchesPattern;
+import static uk.gov.pay.connector.matcher.ResponseContainsLinkMatcher.containsLink;
+import static uk.gov.pay.connector.matcher.ZoneDateTimeAsStringWithinMatcher.isWithin;
+import static uk.gov.pay.connector.model.domain.ChargeStatus.CREATED;
+import static uk.gov.pay.connector.util.JsonEncoder.toJson;
+import static uk.gov.pay.connector.util.NumberMatcher.isNumber;
+
+public class ChargesApiCreateResourceITest extends ChargingITestBase {
+
+    private static final String FRONTEND_CARD_DETAILS_URL = "/secure";
+    private static final String JSON_AMOUNT_KEY = "amount";
+    private static final String JSON_REFERENCE_KEY = "reference";
+    private static final String JSON_DESCRIPTION_KEY = "description";
+    private static final String JSON_GATEWAY_ACC_KEY = "gateway_account_id";
+    private static final String JSON_RETURN_URL_KEY = "return_url";
+    private static final String JSON_CHARGE_KEY = "charge_id";
+    private static final String JSON_STATE_KEY = "state.status";
+    private static final String JSON_MESSAGE_KEY = "message";
+    private static final String JSON_EMAIL_KEY = "email";
+    private static final String JSON_PROVIDER_KEY = "payment_provider";
+    private static final String JSON_LANGUAGE_KEY = "language";
+    private static final String JSON_DELAYED_CAPTURE_KEY = "delayed_capture";
+    private static final String JSON_CORPORATE_SURCHARGE_KEY = "corporate_surcharge";
+    private static final String JSON_TOTAL_AMOUNT_KEY = "total_amount";
+    private static final String PROVIDER_NAME = "sandbox";
+
+    public ChargesApiCreateResourceITest() {
+        super(PROVIDER_NAME);
+    }
+
+    @Test
+    public void makeChargeAndRetrieveAmount() {
+        String expectedReference = "Test reference";
+        String expectedDescription = "Test description";
+        String postBody = toJson(ImmutableMap.builder()
+                .put(JSON_AMOUNT_KEY, AMOUNT)
+                .put(JSON_REFERENCE_KEY, expectedReference)
+                .put(JSON_DESCRIPTION_KEY, expectedDescription)
+                .put(JSON_GATEWAY_ACC_KEY, accountId)
+                .put(JSON_RETURN_URL_KEY, RETURN_URL)
+                .put(JSON_EMAIL_KEY, EMAIL)
+                .put(JSON_LANGUAGE_KEY, "cy")
+                .build()
+        );
+
+        ValidatableResponse response = connectorRestApi
+                .postCreateCharge(postBody)
+                .statusCode(Status.CREATED.getStatusCode())
+                .body(JSON_CHARGE_KEY, is(notNullValue()))
+                .body(JSON_AMOUNT_KEY, isNumber(AMOUNT))
+                .body(JSON_REFERENCE_KEY, is(expectedReference))
+                .body(JSON_DESCRIPTION_KEY, is(expectedDescription))
+                .body(JSON_PROVIDER_KEY, is(PROVIDER_NAME))
+                .body(JSON_RETURN_URL_KEY, is(RETURN_URL))
+                .body(JSON_EMAIL_KEY, is(EMAIL))
+                .body(JSON_LANGUAGE_KEY, is("cy"))
+                .body(JSON_DELAYED_CAPTURE_KEY, is(false))
+                .body("containsKey('card_details')", is(false))
+                .body("containsKey('gateway_account')", is(false))
+                .body("refund_summary.amount_submitted", is(0))
+                .body("refund_summary.amount_available", isNumber(AMOUNT))
+                .body("refund_summary.status", is("pending"))
+                .body("settlement_summary.capture_submit_time", nullValue())
+                .body("settlement_summary.captured_time", nullValue())
+                .body("created_date", matchesPattern("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(.\\d{1,3})?Z"))
+                .body("created_date", isWithin(10, SECONDS))
+                .contentType(JSON);
+
+        String externalChargeId = response.extract().path(JSON_CHARGE_KEY);
+        String documentLocation = expectedChargeLocationFor(accountId, externalChargeId);
+        String chargeTokenId = app.getDatabaseTestHelper().getChargeTokenByExternalChargeId(externalChargeId);
+
+        String hrefNextUrl = "http://Frontend" + FRONTEND_CARD_DETAILS_URL + "/" + chargeTokenId;
+        String hrefNextUrlPost = "http://Frontend" + FRONTEND_CARD_DETAILS_URL;
+
+        response.header("Location", is(documentLocation))
+                .body("links", hasSize(4))
+                .body("links", containsLink("self", "GET", documentLocation))
+                .body("links", containsLink("refunds", "GET", documentLocation + "/refunds"))
+                .body("links", containsLink("next_url", "GET", hrefNextUrl))
+                .body("links", containsLink("next_url_post", "POST", hrefNextUrlPost, "application/x-www-form-urlencoded", new HashMap<String, Object>() {{
+                    put("chargeTokenId", chargeTokenId);
+                }}));
+
+        ValidatableResponse getChargeResponse = connectorRestApi
+                .withAccountId(accountId)
+                .withChargeId(externalChargeId)
+                .getCharge()
+                .statusCode(OK.getStatusCode())
+                .contentType(JSON)
+                .body(JSON_CHARGE_KEY, is(externalChargeId))
+                .body(JSON_AMOUNT_KEY, isNumber(AMOUNT))
+                .body(JSON_REFERENCE_KEY, is(expectedReference))
+                .body(JSON_DESCRIPTION_KEY, is(expectedDescription))
+                .body(JSON_STATE_KEY, is(CREATED.toExternal().getStatus()))
+                .body(JSON_RETURN_URL_KEY, is(RETURN_URL))
+                .body(JSON_EMAIL_KEY, is(EMAIL))
+                .body(JSON_LANGUAGE_KEY, is("cy"))
+                .body(JSON_DELAYED_CAPTURE_KEY, is(false))
+                .body(JSON_CORPORATE_SURCHARGE_KEY, is(nullValue()))
+                .body(JSON_TOTAL_AMOUNT_KEY, is(nullValue()))
+                .body("containsKey('card_details')", is(false))
+                .body("containsKey('gateway_account')", is(false))
+                .body("settlement_summary.capture_submit_time", nullValue())
+                .body("settlement_summary.captured_time", nullValue())
+                .body("refund_summary.amount_submitted", is(0))
+                .body("refund_summary.amount_available", isNumber(AMOUNT))
+                .body("refund_summary.status", is("pending"));
+
+
+        // Reload the charge token which as it should have changed
+        String newChargeTokenId = app.getDatabaseTestHelper().getChargeTokenByExternalChargeId(externalChargeId);
+
+        String newHrefNextUrl = "http://Frontend" + FRONTEND_CARD_DETAILS_URL + "/" + newChargeTokenId;
+
+        getChargeResponse
+                .body("links", hasSize(4))
+                .body("links", containsLink("self", "GET", documentLocation))
+                .body("links", containsLink("refunds", "GET", documentLocation + "/refunds"))
+                .body("links", containsLink("next_url", "GET", newHrefNextUrl))
+                .body("links", containsLink("next_url_post", "POST", hrefNextUrlPost, "application/x-www-form-urlencoded", new HashMap<String, Object>() {{
+                    put("chargeTokenId", newChargeTokenId);
+                }}));
+
+    }
+
+    @Test
+    public void makeChargeWithNoExplicitLanguageDefaultsToEnglish() {
+        String postBody = toJson(ImmutableMap.builder()
+                .put(JSON_AMOUNT_KEY, AMOUNT)
+                .put(JSON_REFERENCE_KEY, "Test reference")
+                .put(JSON_DESCRIPTION_KEY, "Test description")
+                .put(JSON_GATEWAY_ACC_KEY, accountId)
+                .put(JSON_RETURN_URL_KEY, RETURN_URL)
+                .put(JSON_EMAIL_KEY, EMAIL)
+                .build()
+        );
+
+        ValidatableResponse response = connectorRestApi
+                .postCreateCharge(postBody)
+                .statusCode(Status.CREATED.getStatusCode())
+                .body(JSON_LANGUAGE_KEY, is("en"))
+                .contentType(JSON);
+
+        String externalChargeId = response.extract().path(JSON_CHARGE_KEY);
+
+        connectorRestApi
+                .withAccountId(accountId)
+                .withChargeId(externalChargeId)
+                .getCharge()
+                .statusCode(OK.getStatusCode())
+                .contentType(JSON)
+                .body(JSON_LANGUAGE_KEY, is("en"));
+    }
+
+    @Test
+    public void makeChargeNoEmailField_shouldReturnOK() {
+        String expectedReference = "Test reference";
+        String expectedDescription = "Test description";
+        String postBody = toJson(ImmutableMap.builder()
+                .put(JSON_AMOUNT_KEY, AMOUNT)
+                .put(JSON_REFERENCE_KEY, expectedReference)
+                .put(JSON_DESCRIPTION_KEY, expectedDescription)
+                .put(JSON_GATEWAY_ACC_KEY, accountId)
+                .put(JSON_EMAIL_KEY, EMAIL)
+                .put(JSON_RETURN_URL_KEY, RETURN_URL).build());
+
+
+        connectorRestApi
+                .postCreateCharge(postBody)
+                .statusCode(Status.CREATED.getStatusCode())
+                .body(JSON_CHARGE_KEY, is(notNullValue()))
+                .body(JSON_AMOUNT_KEY, isNumber(AMOUNT))
+                .body(JSON_REFERENCE_KEY, is(expectedReference))
+                .body(JSON_DESCRIPTION_KEY, is(expectedDescription))
+                .body(JSON_PROVIDER_KEY, is(PROVIDER_NAME))
+                .body(JSON_RETURN_URL_KEY, is(RETURN_URL))
+                .body("containsKey('card_details')", is(false))
+                .body("containsKey('gateway_account')", is(false))
+                .body("created_date", matchesPattern("^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(.\\d{1,3})?Z"))
+                .body("created_date", isWithin(10, SECONDS))
+                .contentType(JSON);
+
+    }
+
+    @Test
+    public void shouldReturn404WhenCreatingChargeAccountIdIsNonNumeric() {
+
+        String postBody = toJson(ImmutableMap.of(
+                JSON_AMOUNT_KEY, AMOUNT,
+                JSON_REFERENCE_KEY, "Test reference",
+                JSON_DESCRIPTION_KEY, "Test description",
+                JSON_GATEWAY_ACC_KEY, accountId,
+                JSON_RETURN_URL_KEY, RETURN_URL));
+
+        connectorRestApi
+                .withAccountId("invalidAccountId")
+                .postCreateCharge(postBody)
+                .contentType(JSON)
+                .statusCode(NOT_FOUND.getStatusCode())
+                .body("code", is(404))
+                .body("message", is("HTTP 404 Not Found"));
+    }
+    
+    @Test
+    public void shouldReturn400WhenAmountIsLessThanMinAmount() {
+
+        String expectedReference = "Test reference";
+        String expectedDescription = "Test description";
+        String postBody = toJson(ImmutableMap.builder()
+                .put(JSON_AMOUNT_KEY, 0)
+                .put(JSON_REFERENCE_KEY, expectedReference)
+                .put(JSON_DESCRIPTION_KEY, expectedDescription)
+                .put(JSON_GATEWAY_ACC_KEY, accountId)
+                .put(JSON_RETURN_URL_KEY, RETURN_URL)
+                .put(JSON_EMAIL_KEY, EMAIL).build());
+
+        connectorRestApi
+                .postCreateCharge(postBody)
+                .statusCode(Status.BAD_REQUEST.getStatusCode());
+    }
+
+    @Test
+    public void shouldReturn400WhenLanguageNotSupported() {
+        String postBody = toJson(ImmutableMap.builder()
+                .put(JSON_AMOUNT_KEY, AMOUNT)
+                .put(JSON_REFERENCE_KEY, "Test reference")
+                .put(JSON_DESCRIPTION_KEY, "Test description")
+                .put(JSON_GATEWAY_ACC_KEY, accountId)
+                .put(JSON_RETURN_URL_KEY, RETURN_URL)
+                .put(JSON_EMAIL_KEY, EMAIL)
+                .put(JSON_LANGUAGE_KEY, "not a supported language")
+                .build()
+        );
+
+        connectorRestApi
+                .postCreateCharge(postBody)
+                .statusCode(Status.BAD_REQUEST.getStatusCode());
+
+    }
+
+    @Test
+    public void cannotMakeChargeForMissingGatewayAccount() {
+        String missingGatewayAccount = "1234123";
+        String postBody = toJson(ImmutableMap.of(
+                JSON_AMOUNT_KEY, AMOUNT,
+                JSON_REFERENCE_KEY, "Test reference",
+                JSON_DESCRIPTION_KEY, "Test description",
+                JSON_EMAIL_KEY, EMAIL,
+                JSON_RETURN_URL_KEY, RETURN_URL));
+
+        connectorRestApi
+                .withAccountId(missingGatewayAccount)
+                .postCreateCharge(postBody)
+                .statusCode(NOT_FOUND.getStatusCode())
+                .contentType(JSON)
+                .header("Location", is(nullValue()))
+                .body(JSON_CHARGE_KEY, is(nullValue()))
+                .body(JSON_MESSAGE_KEY, is("Unknown gateway account: " + missingGatewayAccount));
+    }
+
+    @Test
+    public void cannotMakeChargeForInvalidSizeOfFields() {
+        String postBody = toJson(ImmutableMap.builder()
+                .put(JSON_AMOUNT_KEY, AMOUNT)
+                .put(JSON_REFERENCE_KEY, randomAlphabetic(256))
+                .put(JSON_DESCRIPTION_KEY, randomAlphanumeric(256))
+                .put(JSON_EMAIL_KEY, randomAlphanumeric(255))
+                .put(JSON_GATEWAY_ACC_KEY, accountId)
+                .put(JSON_RETURN_URL_KEY, RETURN_URL).build());
+
+        connectorRestApi.postCreateCharge(postBody)
+                .statusCode(BAD_REQUEST.getStatusCode())
+                .contentType(JSON)
+                .header("Location", is(nullValue()))
+                .body(JSON_CHARGE_KEY, is(nullValue()))
+                .body(JSON_MESSAGE_KEY, is("Field(s) are too big: [description, reference, email]"));
+    }
+
+    @Test
+    public void cannotMakeChargeForMissingFields() {
+        connectorRestApi.postCreateCharge("{}")
+                .statusCode(BAD_REQUEST.getStatusCode())
+                .contentType(JSON)
+                .header("Location", is(nullValue()))
+                .body(JSON_CHARGE_KEY, is(nullValue()))
+                .body(JSON_MESSAGE_KEY, is("Field(s) missing: [amount, description, reference, return_url]"));
+    }
+    
+    private String expectedChargeLocationFor(String accountId, String chargeId) {
+        return "https://localhost:" + app.getLocalPort() + "/v1/api/accounts/{accountId}/charges/{chargeId}"
+                .replace("{accountId}", accountId)
+                .replace("{chargeId}", chargeId);
+    }
+}
